requirements.txt
python-dotenv
ipython
psycopg2
Scrapy
SQLAlchemy
Twisted
pandas
matplotlib
sklearn
numpy
nltk
Flask
newsdata
